refactor: model to entity
