+++ src/main/java/com/devcard/devcard/card/controller/CardController.java
@@ -1,7 +1,7 @@
 package com.devcard.devcard.card.controller;
 
 import com.devcard.devcard.card.dto.CardResponseDto;
-import com.devcard.devcard.card.model.Card;
+import com.devcard.devcard.card.entity.Card;
 import com.devcard.devcard.card.dto.CardRequestDto;
 import com.devcard.devcard.card.service.CardService;
 import org.springframework.http.ResponseEntity;
