+++ src/main/java/com/devcard/devcard/card/repository/CardRepository.java
@@ -1,6 +1,6 @@
 package com.devcard.devcard.card.repository;
 
-import com.devcard.devcard.card.model.Card;
+import com.devcard.devcard.card.entity.Card;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CardRepository extends JpaRepository<Card, Long>{
